+++ src/main/java/me/pajic/simple_smithing_overhaul/mixin/AbstractContainerMenuMixin.java
@@ -17,6 +17,7 @@ public class AbstractContainerMenuMixin {
 
     @Shadow @Final public NonNullList<Slot> slots;
 
+    @SuppressWarnings({"LocalMayUseName"})
     @ModifyArg(
 			method = "lambda$doClick$0",
             at = @At(