+++ src/main/java/me/pajic/simple_smithing_overhaul/mixin/EnchantmentHelperMixin.java
@@ -19,9 +19,9 @@ public class EnchantmentHelperMixin {
                     target = "Lnet/minecraft/world/item/ItemStack;is(Ljava/lang/Object;)Z"
             )
     )
-    private static boolean storeEnchantmentsIfWhetstone(boolean original, @Local(argsOnly = true) ItemStack stack) {
+    private static boolean storeEnchantmentsIfWhetstone(boolean original, @Local(argsOnly = true, name = "itemStack") ItemStack itemStack) {
         if (SSO.CONFIG.portableItemRepair.enableWhetstone.get()) {
-            return original || stack.is(ModItems.WHETSTONE);
+            return original || itemStack.is(ModItems.WHETSTONE);
         }
         return original;
     }
@@ -33,9 +33,9 @@ public class EnchantmentHelperMixin {
                     target = "Lnet/minecraft/world/item/ItemStack;is(Ljava/lang/Object;)Z"
             )
     )
-    private static boolean handleWhetstoneEnchanting(boolean original, @Local(argsOnly = true) ItemStack stack) {
+    private static boolean handleWhetstoneEnchanting(boolean original, @Local(argsOnly = true, name = "itemStack") ItemStack itemStack) {
         if (SSO.CONFIG.portableItemRepair.enableWhetstone.get()) {
-            return original || stack.is(ModItems.WHETSTONE);
+            return original || itemStack.is(ModItems.WHETSTONE);
         }
         return original;
     }