+++ src/main/java/me/pajic/simple_smithing_overhaul/mixin/client/MinecraftClientMixin.java
@@ -3,7 +3,6 @@ package me.pajic.simple_smithing_overhaul.mixin.client;
 import dev.kikugie.fletching_table.annotation.MixinEnvironment;
 import me.pajic.simple_smithing_overhaul.util.ModUtil;
 import net.minecraft.client.Minecraft;
-import net.minecraft.client.multiplayer.ClientLevel;
 import net.minecraft.core.registries.BuiltInRegistries;
 import net.minecraft.resources.Identifier;
 import org.spongepowered.asm.mixin.Mixin;
@@ -19,7 +18,7 @@ public class MinecraftClientMixin {
 			method = "updateLevelInEngines(Lnet/minecraft/client/multiplayer/ClientLevel;Z)V",
 			at = @At("TAIL")
 	)
-	private void afterClientWorldChange(ClientLevel clientLevel, boolean bl, CallbackInfo ci) {
+	private void afterClientWorldChange( CallbackInfo ci) {
 		ModUtil.itemSuggestions.clear();
 		ModUtil.itemSuggestions.addAll(BuiltInRegistries.ITEM.keySet().stream().map(Identifier::toString).toList());
 		BuiltInRegistries.ITEM.listTagIds().forEach(tag -> ModUtil.itemSuggestions.add("#" + tag.location()));