+++ CHANGELOG.md
@@ -1,2 +1,2 @@
-- Fixed being unable to connect to servers with Simple Smithing Overhaul installed before loading into a singleplayer world first.
-- Mixson is no longer included in the mod and is downloaded separately.
+- The mod now uses Defaulted for creating additional item repairs.
+- Fixed being able to use certain broken items when held in the offhand.